Restructure JavaHelper packages
