+++ src/main/java/tf/gpx/edit/helper/GPXTableView.java
@@ -83,11 +83,11 @@ import tf.gpx.edit.items.GPXTrackSegment;
 import tf.gpx.edit.items.GPXWaypoint;
 import tf.gpx.edit.main.GPXEditor;
 import tf.gpx.edit.main.StatusBar;
-import tf.helper.general.AppClipboard;
+import tf.helper.javafx.AppClipboard;
 import tf.helper.general.ObjectsHelper;
 import tf.helper.javafx.TableMenuUtils;
 import tf.helper.javafx.TooltipHelper;
-import tf.helper.general.UsefulKeyCodes;
+import tf.helper.javafx.UsefulKeyCodes;
 
 /**
  *
