+++ src/main/java/tf/gpx/edit/helper/GPXTreeTableView.java
@@ -90,13 +90,13 @@ import tf.gpx.edit.items.GPXWaypoint;
 import tf.gpx.edit.main.GPXEditor;
 import tf.gpx.edit.srtm.SRTMDataViewer;
 import tf.gpx.edit.viewer.GPXTrackviewer;
-import tf.helper.general.AppClipboard;
-import tf.helper.general.ColorConverter;
+import tf.helper.javafx.AppClipboard;
+import tf.helper.javafx.ColorConverter;
 import tf.helper.javafx.ColorSelectionMenu;
 import tf.helper.general.ObjectsHelper;
 import tf.helper.javafx.TableMenuUtils;
 import tf.helper.javafx.TooltipHelper;
-import tf.helper.general.UsefulKeyCodes;
+import tf.helper.javafx.UsefulKeyCodes;
 
 /**
  *
